common: move *.h from battle_interface/src to include folder
