+++ src/libs/battle_interface/include/battle_sign.h
