+++ src/libs/battle_interface/include/bi_defines.h
