+++ src/libs/battle_interface/include/utils.h
