+++ src/libs/battle_interface/src/interface_manager/interface_group/sea_group.cpp
@@ -1,5 +1,5 @@
 #include "sea_group.h"
-#include "../../utils.h"
+#include "utils.h"
 #include "platform/platform.hpp"
 
 BI_SeaGroup::BI_SeaGroup(BI_ManagerBase *pManager) : BI_BaseGroup(pManager)
