+++ src/libs/battle_interface/src/interface_manager/interface_manager.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../bi_defines.h"
+#include "bi_defines.h"
 #include "base_manager.h"
 #include <vector>
