+++ src/libs/battle_interface/src/interface_manager/interface_node/image_node.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../bi_defines.h"
+#include "bi_defines.h"
 #include "base_node.h"
 
 class IBIImage;
