+++ src/libs/battle_interface/src/interface_manager/mouse_pointer.cpp
@@ -1,5 +1,5 @@
 #include "mouse_pointer.h"
-#include "../utils.h"
+#include "utils.h"
 #include "../image/img_render.h"
 #include "core.h"
 #include "vma.hpp"
