+++ src/libs/battle_interface/src/land/battle_land.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "../utils.h"
-#include "../bi_defines.h"
+#include "utils.h"
+#include "bi_defines.h"
 
 class BIManSign;
