+++ src/libs/battle_interface/src/land/battle_man_sign.cpp
@@ -1,5 +1,5 @@
 #include "battle_man_sign.h"
-#include "../utils.h"
+#include "utils.h"
 #include "battle_man_command.h"
 #include "core.h"
 #include "shared/battle_interface/msg_control.h"
