+++ src/libs/battle_interface/src/land/i_boarding_status.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../bi_defines.h"
+#include "bi_defines.h"
 
 class IBoardingStatus : public Entity
 {
