+++ src/libs/battle_interface/src/sea/i_battle.cpp
@@ -13,9 +13,9 @@
 #include "../land/i_boarding_status.h"
 #include "../land/battle_land.h"
 
-#include "../active_perk_shower.h"
+#include "active_perk_shower.h"
+#include "log_and_action.h"
 #include "../item_entity/item_entity.h"
-#include "../log_and_action.h"
 #include "../world_map_interface/interface.h"
 #include "../spyglass/spyglass.h"
 #include "../timer/timer.h"
