+++ src/libs/battle_interface/src/sea/i_battle.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "../utils.h"
-#include "../bi_defines.h"
+#include "utils.h"
+#include "bi_defines.h"
 #include "battle_navigator.h"
 #include "dx9render.h"
