+++ src/libs/battle_interface/src/timer/timer.cpp
@@ -1,5 +1,5 @@
 #include "timer.h"
-#include "../utils.h"
+#include "utils.h"
 #include "../image/img_render.h"
 #include "core.h"
 #include "message.h"
