+++ src/libs/battle_interface/src/world_map_interface/ship_command.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../battle_command.h"
+#include "battle_command.h"
 
 class WMShipCommandList : public BICommandList
 {
