+++ src/libs/battle_interface/src/world_map_interface/ship_sign.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "../battle_sign.h"
-#include "../bi_defines.h"
+#include "battle_sign.h"
+#include "bi_defines.h"
 
 #define MAX_SHIP_QUANTITY 8
